requirements.txt
alembic==0.9.7
attrdict==2.0.1
beautifulsoup4==4.7.1
Bottleneck==1.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
dill==0.2.9
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
eth-account==0.2.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.1
eth-rlp==0.1.2
eth-typing==2.0.0
eth-utils==1.4.1
fastai==1.0.43.dev0
fastprogress==0.1.18
hexbytes==0.1.0
idna==2.7
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.2
mkl-fft==1.0.10
mkl-random==1.0.2
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
numexpr==2.6.4
numpy==1.16.1
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==19.0
pandas==0.24.1
Pillow==5.4.1
plac==0.9.6
preshed==2.0.1
psycopg2==2.7.7
pycparser==2.19
pycryptodome==3.7.3
pyparsing==2.3.1
python-dateutil==2.7.3
python-editor==1.0.4
pytz==2018.9
PyYAML==3.13
regex==2018.1.10
requests==2.21.0
rlp==1.1.0
scipy==1.2.0
six==1.11.0
soupsieve==1.7.3
spacy==2.0.18
SQLAlchemy==1.2.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tqdm==4.30.0
typing==3.6.6
ujson==1.35
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
